requirements.txt
asgiref==3.8.1
Django==5.1.7
django-ckeditor==6.7.2
django-js-asset==3.1.2
django-modeltranslation==0.19.13
django-widget-tweaks==1.5.0
gunicorn==23.0.0
joblib==1.4.2
mysqlclient==2.2.7
numpy==2.2.4
packaging==24.2
pillow==11.1.0
scikit-learn==1.6.1
scipy==1.15.2
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.12.2
